Remove duplicate exceptions file
