+++ trailblazer/services/slurm/slurm_api_service/error_handler.py
@@ -1,0 +1,22 @@
+import logging
+from functools import wraps
+
+from trailblazer.exc import TowerAPIClientError
+from trailblazer.exc import TowerServiceError, TowerWorkflowIdFileMissing
+
+LOG = logging.getLogger(__name__)
+
+
+def handle_tower_service_errors(func):
+
+    @wraps(func)
+    def wrapper(*args, **kwargs):
+        try:
+            return func(*args, **kwargs)
+        except FileNotFoundError as error:
+            LOG.error(f"Job ids file not found in slurm service: {error}")
+            raise TowerWorkflowIdFileMissing()
+        except TowerAPIClientError as error:
+            raise TowerServiceError("Error in tower API client")
+
+    return wrapper
